Add files via upload
The goal of this model is to predict whether an existing customer will buy a product in online based on the two input features: Age and current Salary per month. The dataset consist these features. *
The model used in this proect is logistic regression, a popular method for binary classification tasks.
Logistic regression is the classification model uses statistical algorithm tht models the relationship between the input variables and the binary outcome, in this case the output will be whether the customer will buy this product in online through the advertisement or not.
The model uses the customer's age and current salary per month as
predictors(features\independent variables) to make this predicton by estimating the probablities of the binary outcome using a logistic function.
Once the model is trained, it can be used to make predicitons on new, unseen data. Given the age and monthly salary of the new customers, the logistic regression model can predict the likelyhood of that customer buying a product online through clicking the advertisement